review crud add status_code & verify_authorization_feature

+++ server/router/oauth.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter
-from fastapi import Depends, FastAPI, HTTPException
+from fastapi import Depends, FastAPI, HTTPException, status, Response
 from starlette.config import Config
 from fastapi.responses import JSONResponse
 from sqlalchemy.orm import Session, sessionmaker
@@ -95,23 +95,30 @@ async def signUpKakao(code:str, db: Session = Depends(get_db)):
 #     return JSONResponse(resp.json())
 
 
-@router.get("/user/me")
-async def getMyInfo(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
-
-    email = auth.verify_jwttoken(token)
-
-    return crud.get_user_by_email(db= db, email = email)
+@router.get("/user/me", status_code=200)
+async def getMyInfo(response: Response, token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
+    if(auth.verify_jwttoken(token)):
+        email = auth.verify_jwttoken(token)
+        return crud.get_user_by_email(db= db, email = email)
+    else:
+        response.status_code= status_code.HTTP_401_UNAUTHORIZED
+        return {"msg": "인증/재인증이 필요합니다."}
 
 @router.delete("/user/delete", status_code=201)
 async def deleteUser(token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
 
-    email = auth.verify_jwttoken(token)
-
-    return JSONResponse(content= crud.delete_user(db = db, email = email))
-
+    if(auth.verify_jwttoken(token)):
+        email = auth.verify_jwttoken(token)
+        return JSONResponse(content= crud.delete_user(db = db, email = email))
+    else:
+        response.status_code = status_code.HTTP_401_UNAUTHORIZED
+        return {"msg" : "인증/재인증이 필요합니다."}
 @router.post("/user/update/nickname", status_code=200)
 async def updateNickname(nickname:str, token: str = Depends(oauth2_scheme), db: Session = Depends(get_db)):
-
-    email = auth.verify_jwttoken(token)
-    result = crud.update_user_nickname(db = db, email = email, nickname = nickname)
-    return JSONResponse(content = result, status_code=200)
+    if(auth.verify_jwttoken(token)):
+        email = auth.verify_jwttoken(token)
+        result = crud.update_user_nickname(db = db, email = email, nickname = nickname)
+        return JSONResponse(content = result, status_code=200)
+    else:
+        auth.status_code = status_code.HTTP_401_UNAUTHORIZED
+        return {"msg": "인증/재인증이 필요합니다."}
